Bitmap #1

+++ nmm/src/arch/mod.rs
@@ -36,7 +36,7 @@ pub const PHYSICAL_ADDRESS_MAX: u64 = arch_impl::PHYSICAL_ADDRESS_MAX;
 /// The number of bits used for indexing into page tables at each level.
 pub const TABLE_INDEX_BITS: usize = arch_impl::TABLE_INDEX_BITS;
 /// The size of a page table in bytes for x86_64 architecture.
-pub const TABLE_SIZE: usize = arch_impl::TABLE_SIZE;
+pub const TABLE_SIZE: u64 = arch_impl::TABLE_SIZE as u64;
 /// The number of entries in a page table for x86_64 architecture.
 pub const ENTRY_COUNT: usize = arch_impl::ENTRY_COUNT;
